Fix invalid type crash when attempting to recover keyboard landscape height.
A bug with setter using long means it easier to just use long going
forward.

+++ app/src/main/java/org/thoughtcrime/securesms/keyvalue/MiscellaneousValues.java
@@ -262,7 +262,7 @@ public final class MiscellaneousValues extends SignalStoreValues {
   }
 
   public int getKeyboardLandscapeHeight() {
-    int height = getInteger(KEYBOARD_LANDSCAPE_HEIGHT, 0);
+    int height = (int) getLong(KEYBOARD_LANDSCAPE_HEIGHT, 0);
     if (height == 0) {
       //noinspection deprecation
       height = PreferenceManager.getDefaultSharedPreferences(ApplicationDependencies.getApplication())
